+++ Commands/Provisioning/Site/AddFileToSiteTemplate.cs
@@ -11,31 +11,32 @@ using System.Management.Automation;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SharePointPnP.PowerShell.Commands.Provisioning
+namespace SharePointPnP.PowerShell.Commands.Provisioning.Site
 {
-    [Cmdlet(VerbsCommon.Add, "PnPFileToProvisioningTemplate")]
-    [CmdletHelp("Adds a file to a PnP Provisioning Template",
+    [Cmdlet(VerbsCommon.Add, "PnPFileToSiteTemplate")]
+    [Alias("Add-PnPFileToProvisioningTemplate")]
+    [CmdletHelp("Adds a file to a PnP Site Template",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPFileToProvisioningTemplate -Path template.pnp -Source $sourceFilePath -Folder $targetFolder",
-       Remarks = "Adds a file to a PnP Provisioning Template",
+       Code = @"PS:> Add-PnPFileToSiteTemplate -Path template.pnp -Source $sourceFilePath -Folder $targetFolder",
+       Remarks = "Adds a file to a PnP Site Template",
        SortOrder = 1)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPFileToProvisioningTemplate -Path template.xml -Source $sourceFilePath -Folder $targetFolder",
-       Remarks = "Adds a file reference to a PnP Provisioning XML Template",
+       Code = @"PS:> Add-PnPFileToSiteTemplate -Path template.xml -Source $sourceFilePath -Folder $targetFolder",
+       Remarks = "Adds a file reference to a PnP Site XML Template",
        SortOrder = 2)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPFileToProvisioningTemplate -Path template.pnp -Source ""./myfile.png"" -Folder ""folderinsite"" -FileLevel Published -FileOverwrite:$false",
-       Remarks = "Adds a file to a PnP Provisioning Template, specifies the level as Published and defines to not overwrite the file if it exists in the site.",
+       Code = @"PS:> Add-PnPFileToSiteTemplate -Path template.pnp -Source ""./myfile.png"" -Folder ""folderinsite"" -FileLevel Published -FileOverwrite:$false",
+       Remarks = "Adds a file to a PnP Site Template, specifies the level as Published and defines to not overwrite the file if it exists in the site.",
        SortOrder = 3)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPFileToProvisioningTemplate -Path template.pnp -Source $sourceFilePath -Folder $targetFolder -Container $container",
-       Remarks = "Adds a file to a PnP Provisioning Template with a custom container for the file",
+       Code = @"PS:> Add-PnPFileToSiteTemplate -Path template.pnp -Source $sourceFilePath -Folder $targetFolder -Container $container",
+       Remarks = "Adds a file to a PnP Site Template with a custom container for the file",
        SortOrder = 4)]
 
     public class AddFileToProvisioningTemplate : PSCmdlet
     {
-        [Parameter(Mandatory = true, Position = 0, HelpMessage = "Filename of the .PNP Open XML provisioning template to read from, optionally including full path.")]
+        [Parameter(Mandatory = true, Position = 0, HelpMessage = "Filename of the .PNP Open XML site template to read from, optionally including full path.")]
         public string Path;
 
         [Parameter(Mandatory = true, Position = 1, HelpMessage = "The file to add to the in-memory template, optionally including full path.")]
@@ -67,7 +68,7 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
                 Source = System.IO.Path.Combine(SessionState.Path.CurrentFileSystemLocation.Path, Source);
             }
             // Load the template
-            var template = ReadProvisioningTemplate
+            var template = ReadSiteTemplate
                 .LoadProvisioningTemplateFromFile(Path,
                 TemplateProviderExtensions);