+++ Commands/Provisioning/Site/NewSiteTemplate.cs
@@ -4,14 +4,15 @@ using System.Management.Automation;
 
 namespace SharePointPnP.PowerShell.Commands.Provisioning
 {
-    [Cmdlet(VerbsCommon.New, "PnPProvisioningTemplate", SupportsShouldProcess = true)]
-    [CmdletHelp("Creates a new provisioning template object",
+    [Cmdlet(VerbsCommon.New, "PnPSiteTemplate", SupportsShouldProcess = true)]
+    [Alias("New-PnPProvisioningTemplate")]
+    [CmdletHelp("Creates a new site template object",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
-       Code = @"PS:> $template = New-PnPProvisioningTemplate",
-       Remarks = "Creates a new instance of a provisioning template object.",
+       Code = @"PS:> $template = New-PnPSiteTemplate",
+       Remarks = "Creates a new instance of a site template object.",
        SortOrder = 1)]
-    public class NewProvisioningTemplate : PSCmdlet
+    public class NewSiteTemplate : PSCmdlet
     {
         protected override void ProcessRecord()
         {