+++ Commands/Provisioning/Tenant/AddTenantSequenceSite.cs
@@ -4,16 +4,17 @@ using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 using System.Management.Automation;
 
-namespace SharePointPnP.PowerShell.Commands.Provisioning
+namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
 {
-    [Cmdlet(VerbsCommon.Add, "PnPProvisioningSite", SupportsShouldProcess = true)]
-    [CmdletHelp("Adds a provisioning sequence object to a provisioning hierarchy",
+    [Cmdlet(VerbsCommon.Add, "PnPTenantSequenceSite", SupportsShouldProcess = true)]
+    [Alias("Add-PnPProvisioningSite")]
+    [CmdletHelp("Adds a existing tenant sequence site object to a tenant template",
         Category = CmdletHelpCategory.Provisioning, SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPProvisioningSite -Site $myteamsite -Sequence $mysequence",
+       Code = @"PS:> Add-PnPTenantSequenceSite -Site $myteamsite -Sequence $mysequence",
        Remarks = "Adds an existing site object to an existing hierarchy sequence",
         SortOrder = 1)]
-    public class AddProvisioningSite : PSCmdlet
+    public class AddTenantSequenceSite : PSCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true)]
         public ProvisioningSitePipeBind Site;