+++ Commands/Provisioning/Tenant/AddTenantSequenceSubSite.cs
@@ -5,16 +5,17 @@ using System;
 using System.Linq;
 using System.Management.Automation;
 
-namespace SharePointPnP.PowerShell.Commands.Provisioning
+namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
 {
-    [Cmdlet(VerbsCommon.Add, "PnPProvisioningSubSite", SupportsShouldProcess = true)]
-    [CmdletHelp("Adds a provisioning sequence object to a provisioning site object",
+    [Cmdlet(VerbsCommon.Add, "PnPTenantSequenceSubSite", SupportsShouldProcess = true)]
+    [Alias("Add-PnPPnPProvisioningSubSite")]
+    [CmdletHelp("Adds a tenant sequence sub site object to a tenant sequence site object",
         Category = CmdletHelpCategory.Provisioning, SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-       Code = @"PS:> Add-PnPProvisioningSubSite -Site $mysite -SubSite $mysubsite",
-       Remarks = "Adds an existing subsite object to an existing hierarchy sequence site object",
+       Code = @"PS:> Add-PnPTenantSequenceSubSite -Site $mysite -SubSite $mysubsite",
+       Remarks = "Adds an existing subsite object to an existing sequence site object",
        SortOrder = 1)]
-    public class AddProvisioningSubSite : PSCmdlet
+    public class AddTenantSequenceSubSite : PSCmdlet
     {
         [Parameter(Mandatory = true, HelpMessage = "The subsite to add")]
         public TeamNoGroupSubSite SubSite;