+++ Commands/Provisioning/Tenant/GetTenantSequence.cs
@@ -4,23 +4,24 @@ using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 using System;
 using System.Management.Automation;
 
-namespace SharePointPnP.PowerShell.Commands.Provisioning
+namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
 {
-    [Cmdlet(VerbsCommon.Get, "PnPProvisioningSequence", SupportsShouldProcess = true)]
-    [CmdletHelp("Returns one ore more provisioning sequence object(s) from a provisioning hierarchy",
+    [Cmdlet(VerbsCommon.Get, "PnPTenantSequence", SupportsShouldProcess = true)]
+    [Alias("Get-PnPProvisioningSequence")]
+    [CmdletHelp("Returns one ore more provisioning sequence object(s) from a tenant template",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
-       Code = @"PS:> Get-PnPProvisioningSequence -Hierarchy $myhierarchy",
-       Remarks = "Returns all sequences from the specified hierarchy",
+       Code = @"PS:> Get-PnPTenantSequence -Template $myhierarchy",
+       Remarks = "Returns all sequences from the specified tenant template",
        SortOrder = 1)]
     [CmdletExample(
-       Code = @"PS:> Get-PnPProvisioningSequence -Hierarchy $myhierarchy -Identity ""mysequence""",
-       Remarks = "Returns the specified sequence from the specified hierarchy",
+       Code = @"PS:> Get-PnPTenantSequence -Template $myhierarchy -Identity ""mysequence""",
+       Remarks = "Returns the specified sequence from the specified tenant template",
        SortOrder = 2)]
-    public class GetProvisioningSequence : PSCmdlet
+    public class GetTenantSequence : PSCmdlet
     {
         [Parameter(Mandatory = true, HelpMessage = "The hierarchy to retrieve the sequence from", ParameterSetName = ParameterAttribute.AllParameterSets)]
-        public ProvisioningHierarchy Hierarchy;
+        public ProvisioningHierarchy Template;
 
         [Parameter(Mandatory = false, HelpMessage = "Optional Id of the sequence", ParameterSetName = ParameterAttribute.AllParameterSets, ValueFromPipeline = true)]
         public ProvisioningSequencePipeBind Identity;
@@ -28,11 +29,11 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
         {
             if (!MyInvocation.BoundParameters.ContainsKey("Identity"))
             {
-                WriteObject(Hierarchy.Sequences, true);
+                WriteObject(Template.Sequences, true);
             }
             else
             {
-                WriteObject(Identity.GetSequenceFromHierarchy(Hierarchy));
+                WriteObject(Identity.GetSequenceFromHierarchy(Template));
             }
         }
     }