+++ Commands/Provisioning/Tenant/NewTenantSequence.cs
@@ -4,20 +4,21 @@ using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using System;
 using System.Management.Automation;
 
-namespace SharePointPnP.PowerShell.Commands.Provisioning
+namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
 {
-    [Cmdlet(VerbsCommon.New, "PnPProvisioningSequence", SupportsShouldProcess = true)]
-    [CmdletHelp("Creates a new provisioning sequence object",
+    [Cmdlet(VerbsCommon.New, "PnPTenantSequence", SupportsShouldProcess = true)]
+    [Alias("New-PnPProvisioningSequence")]
+    [CmdletHelp("Creates a new tenant sequence object",
         Category = CmdletHelpCategory.Provisioning, SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-       Code = @"PS:> $sequence = New-PnPProvisioningSequence",
-       Remarks = "Creates a new instance of a provisioning sequence object.",
+       Code = @"PS:> $sequence = New-PnPTenantSequence",
+       Remarks = "Creates a new instance of a tenant sequence object.",
        SortOrder = 1)]
     [CmdletExample(
-       Code = @"PS:> $sequence = New-PnPProvisioningSequence -Id ""MySequence""",
-       Remarks = "Creates a new instance of a provisioning sequence object and sets the Id to the value specified.",
+       Code = @"PS:> $sequence = New-PnPTenantSequence -Id ""MySequence""",
+       Remarks = "Creates a new instance of a tenant sequence object and sets the Id to the value specified.",
        SortOrder = 2)]
-    public class NewProvisioningSequence : PSCmdlet
+    public class NewTenantSequence : PSCmdlet
     {
         [Parameter(Mandatory = false, HelpMessage = "Optional Id of the sequence", ParameterSetName = ParameterAttribute.AllParameterSets)]
         public string Id;